pyproject.toml
[project]
name = "presidio-anonymization-tool"
version = "0.1.0"
description = "Cross-platform document anonymization tool using Microsoft Presidio"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Albert"}
]
dependencies = [
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "spacy>=3.7.0",
    "python-docx>=1.1.0",
    "PyMuPDF>=1.23.0",
    "reportlab>=4.0.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]
build = [
    "pyinstaller>=6.0.0",
]

[project.scripts]
anonymize = "anonymizer.ports.cli:app"
anonymize-gui = "anonymizer.ports.gui:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/anonymizer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "basic"
